platform update

+++ internal/COMPILER.h
@@ -142,7 +142,7 @@
       typedef unsigned char uint8_t;
       typedef unsigned short uint16_t;
       typedef unsigned int uint32_t;
-      typedef uintptr_t uint64_t;
+      typedef unsigned long long uint64_t;
     #endif
   #elif SYSTEM_BIT == 32
     typedef signed long sintptr_t;
@@ -154,7 +154,7 @@
       typedef unsigned long uintptr_t;
       typedef unsigned char uint8_t;
       typedef unsigned short uint16_t;
-      typedef uintptr_t uint32_t;
+      typedef unsigned int uint32_t;
       typedef unsigned long long uint64_t;
     #endif
   #else
